Move builder -> factory, remove config
