+++ controllers/profiles/dev/profile_dev.go
@@ -15,7 +15,6 @@
 package dev
 
 import (
-	"k8s.io/client-go/rest"
 	"k8s.io/klog/v2"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -36,7 +35,7 @@ func (d developmentProfile) GetProfile() metadata.ProfileType {
 	return metadata.DevProfile
 }
 
-func NewProfileReconciler(client client.Client, config *rest.Config) profiles.ProfileReconciler {
+func NewProfileReconciler(client client.Client) profiles.ProfileReconciler {
 	support := &common.StateSupport{
 		C: client,
 	}
